Apply mentor feedback and add optional task 3
Use float for salaries and keep cat age as string per feedback.
Add task 3: recursive colored directory tree visualizer using
colorama, with its own venv and requirements.txt.

+++ task_1.py
@@ -13,7 +13,7 @@ def total_salary(path):
                     continue
                 try:
                     name, salary = line.split(",")
-                    total += int(salary)
+                    total += float(salary)
                     count += 1
                 except ValueError:
                     print(f"Пропущено некоректний рядок: '{line}'")
